+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 1.0.xx
+
+ - Fixes bug when training model on a single dataset.
+ - Adds tests for a subset of the CLI tools.
+
 ## 1.0.1
 
  - Fixed model name.